java/iot3mobility: rename and relocate DENM PathPosition and PathElement classes for v1.1.3 and v2.2.0
