+++ java/iot3/mobility/src/main/java/com/orange/iot3mobility/messages/denm/v220/codec/DenmReader220.java
@@ -12,8 +12,8 @@ import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonToken;
 import com.orange.iot3mobility.messages.denm.v220.model.DenmEnvelope220;
 import com.orange.iot3mobility.messages.denm.v220.model.DenmMessage220;
-import com.orange.iot3mobility.messages.denm.v220.model.DenmPathElement;
-import com.orange.iot3mobility.messages.denm.v220.model.DenmPathPosition;
+import com.orange.iot3mobility.messages.denm.v220.model.path.PathElement;
+import com.orange.iot3mobility.messages.denm.v220.model.path.PathPosition;
 import com.orange.iot3mobility.messages.denm.v220.model.alacartecontainer.AlacarteContainer;
 import com.orange.iot3mobility.messages.denm.v220.model.defs.Altitude;
 import com.orange.iot3mobility.messages.denm.v220.model.defs.DeltaReferencePosition;
@@ -53,7 +53,7 @@ public final class DenmReader220 {
             String version = null;
             String sourceUuid = null;
             Long timestamp = null;
-            List<DenmPathElement> path = null;
+            List<PathElement> path = null;
             DenmMessage220 message = null;
 
             while (parser.nextToken() != JsonToken.END_OBJECT) {
@@ -83,18 +83,18 @@ public final class DenmReader220 {
         }
     }
 
-    private List<DenmPathElement> readPath(JsonParser parser) throws IOException {
+    private List<PathElement> readPath(JsonParser parser) throws IOException {
         expect(parser.getCurrentToken(), JsonToken.START_ARRAY);
-        List<DenmPathElement> elements = new ArrayList<>();
+        List<PathElement> elements = new ArrayList<>();
         while (parser.nextToken() != JsonToken.END_ARRAY) {
             elements.add(readPathElement(parser));
         }
         return elements;
     }
 
-    private DenmPathElement readPathElement(JsonParser parser) throws IOException {
+    private PathElement readPathElement(JsonParser parser) throws IOException {
         expect(parser.getCurrentToken(), JsonToken.START_OBJECT);
-        DenmPathPosition position = null;
+        PathPosition position = null;
         String messageType = null;
 
         while (parser.nextToken() != JsonToken.END_OBJECT) {
@@ -107,12 +107,12 @@ public final class DenmReader220 {
             }
         }
 
-        return new DenmPathElement(
+        return new PathElement(
                 requireField(position, "path.position"),
                 requireField(messageType, "path.message_type"));
     }
 
-    private DenmPathPosition readPathPosition(JsonParser parser) throws IOException {
+    private PathPosition readPathPosition(JsonParser parser) throws IOException {
         expect(parser.getCurrentToken(), JsonToken.START_OBJECT);
         Integer latitude = null;
         Integer longitude = null;
@@ -129,7 +129,7 @@ public final class DenmReader220 {
             }
         }
 
-        return new DenmPathPosition(
+        return new PathPosition(
                 requireField(latitude, "path.position.latitude"),
                 requireField(longitude, "path.position.longitude"),
                 requireField(altitude, "path.position.altitude"));
